requirements.txt
ultralytics
cv2-python